# Perfomance improvement for flex access/egress searches