--- a/src/ext/java/org/opentripplanner/ext/flex/FlexParameters.java
+++ b/src/ext/java/org/opentripplanner/ext/flex/FlexParameters.java
@@ -1,0 +1,9 @@
+package org.opentripplanner.ext.flex;
+
+public class FlexParameters {
+  public final double maxTransferMeters;
+
+  public FlexParameters(double maxTransferMeters) {
+    this.maxTransferMeters = maxTransferMeters;
+  }
+}