--- a/src/ext/java/org/opentripplanner/ext/flex/FlexRouter.java
+++ b/src/ext/java/org/opentripplanner/ext/flex/FlexRouter.java
@@ -28,11 +28,13 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.opentripplanner.standalone.config.FlexConfig;
 
 public class FlexRouter {
 
   /* Transit data */
   private final Graph graph;
+  private final FlexConfig config;
   private final Collection<NearbyStop> streetAccesses;
   private final Collection<NearbyStop> streetEgresses;
   private final FlexIndex flexIndex;
@@ -52,6 +54,7 @@ public class FlexRouter {
 
   public FlexRouter(
       Graph graph,
+      FlexConfig config,
       Instant searchInstant,
       boolean arriveBy,
       int additionalPastSearchDays,
@@ -60,6 +63,7 @@ public class FlexRouter {
       Collection<NearbyStop> egressTransfers
   ) {
     this.graph = graph;
+    this.config = config;
     this.streetAccesses = streetAccesses;
     this.streetEgresses = egressTransfers;
     this.flexIndex = graph.index.getFlexIndex();
@@ -145,7 +149,8 @@ public class FlexRouter {
             .flatMap(date -> t2.second.getFlexAccessTemplates(
                 t2.first,
                 date,
-                accessFlexPathCalculator
+                accessFlexPathCalculator,
+                config
             )))
         .collect(Collectors.toList());
   }
@@ -163,7 +168,8 @@ public class FlexRouter {
             .flatMap(date -> t2.second.getFlexEgressTemplates(
                 t2.first, 
                 date,
-                egressFlexPathCalculator
+                egressFlexPathCalculator,
+                config
             )))
         .collect(Collectors.toList());;
   }