--- a/src/ext/java/org/opentripplanner/ext/flex/edgetype/FlexTripEdge.java
+++ b/src/ext/java/org/opentripplanner/ext/flex/edgetype/FlexTripEdge.java
@@ -48,7 +48,8 @@ public class FlexTripEdge extends Edge {
   @Override
   public State traverse(State s0) {
     if(this.flexPath == null) {
-      return null; // not routable
+       // not routable
+       return null;
     }
     StateEditor editor = s0.edit(this);
     editor.setBackMode(TraverseMode.BUS);