--- a/src/ext/java/org/opentripplanner/ext/flex/flexpathcalculator/FlexPath.java
+++ b/src/ext/java/org/opentripplanner/ext/flex/flexpathcalculator/FlexPath.java
@@ -1,5 +1,6 @@
 package org.opentripplanner.ext.flex.flexpathcalculator;
 
+import java.util.function.Supplier;
 import org.locationtech.jts.geom.LineString;
 
 /**
@@ -9,11 +10,32 @@ public class FlexPath {
 
   public int distanceMeters;
   public int durationSeconds;
-  public LineString geometry;
+  private final Supplier<LineString> geometrySupplier;
+  private LineString geometry;
 
-  public FlexPath(int distanceMeters, int durationSeconds, LineString geometry) {
+  public FlexPath(int distanceMeters, int durationSeconds, FlexPath path) {
+    this(distanceMeters, durationSeconds, path::getGeometry);
+  }
+
+  /**
+   *
+   * @param geometrySupplier Computing a linestring from a GraphPath is a surprisingly expensive
+   *                         operation and since there are very many instances of these
+   *                         for a flex access/egress search the actual computation is delayed until
+   *                         the linestring is actually needed.
+   *                         Most of them are _not_ needed so this increases performance quite
+   *                         dramatically.
+   */
+  public FlexPath(int distanceMeters, int durationSeconds, Supplier<LineString> geometrySupplier) {
     this.distanceMeters = distanceMeters;
     this.durationSeconds = durationSeconds;
-    this.geometry = geometry;
+    this.geometrySupplier = geometrySupplier;
+  }
+
+  public LineString getGeometry() {
+    if(geometry == null) {
+      geometry = geometrySupplier.get();
+    }
+    return geometry;
   }
 }