--- a/src/ext/java/org/opentripplanner/ext/flex/template/FlexAccessTemplate.java
+++ b/src/ext/java/org/opentripplanner/ext/flex/template/FlexAccessTemplate.java
@@ -21,13 +21,13 @@ import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.graphfinder.NearbyStop;
 import org.opentripplanner.routing.spt.GraphPath;
-import org.opentripplanner.ext.flex.FlexConfig;
+import org.opentripplanner.ext.flex.FlexParameters;
 
 public class FlexAccessTemplate extends FlexAccessEgressTemplate {
   public FlexAccessTemplate(
       NearbyStop accessEgress, FlexTrip trip, int fromStopTime, int toStopTime,
       StopLocation transferStop, FlexServiceDate date, FlexPathCalculator calculator,
-      FlexConfig config
+      FlexParameters config
   ) {
     super(accessEgress, trip, fromStopTime, toStopTime, transferStop, date, calculator, config);
   }