--- a/src/ext/java/org/opentripplanner/ext/flex/template/FlexEgressTemplate.java
+++ b/src/ext/java/org/opentripplanner/ext/flex/template/FlexEgressTemplate.java
@@ -15,13 +15,13 @@ import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.graphfinder.NearbyStop;
-import org.opentripplanner.ext.flex.FlexConfig;
+import org.opentripplanner.ext.flex.FlexParameters;
 
 public class FlexEgressTemplate extends FlexAccessEgressTemplate {
   public FlexEgressTemplate(
       NearbyStop accessEgress, FlexTrip trip, int fromStopTime, int toStopTime,
       StopLocation transferStop, FlexServiceDate date, FlexPathCalculator calculator,
-      FlexConfig config
+      FlexParameters config
   ) {
     super(accessEgress, trip, fromStopTime, toStopTime, transferStop, date, calculator, config);
   }