--- a/src/ext/java/org/opentripplanner/ext/flex/trip/ScheduledDeviatedTrip.java
+++ b/src/ext/java/org/opentripplanner/ext/flex/trip/ScheduledDeviatedTrip.java
@@ -118,8 +118,8 @@ public class ScheduledDeviatedTrip extends FlexTrip {
                 stop,
                 date,
                 scheduledCalculator,
-                config)
-        );
+                config
+        ));
       }
     }