### Summary

We noticed a severe performance problem for flex access/egress searches which would typically be 5-10 times slower than flex direct searches. I also saw that for flex access/egress searches a large number (25000 in my simple case with a hand full of flex routes) of accesses are computed. For this reason I added their number to the debug output.

After a bit of investigation I noticed some performance improvements in #3499 which I extracted into this separate PR. The core of this solution is to stop every flex path from being traversed twice.

#### Results

Before:
```
10:20:25.981 DEBUG (DebugTimingAggregator.java:220) ┌  Routing initialization           :     0 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) ├  Direct street routing            :     6 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) │┌ Creating raptor data model       :    30 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) │├ Access routing (25105 accesses)  :  4382 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) │├ Egress routing (432 egresses)    :    83 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) ││ Access/Egress routing            :  4465 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) │├ Main routing                     :   477 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) │├ Creating itineraries             :    10 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) ├┴ Transit routing total            :  4982 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) │  Routing total:                   :  4988 ms
10:20:25.982 DEBUG (DebugTimingAggregator.java:220) ├  Filtering itineraries            :     1 ms
10:20:25.983 DEBUG (DebugTimingAggregator.java:220) ├  Converting model objects         :     0 ms
10:20:25.983 DEBUG (DebugTimingAggregator.java:220) ┴  Request total                    :  4989 ms
```

After:

```
10:10:48.699 DEBUG (DebugTimingAggregator.java:220) ┌  Routing initialization           :     0 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) ├  Direct street routing            :     6 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) │┌ Creating raptor data model       :    21 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) │├ Access routing (25117 accesses)  :  2745 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) │├ Egress routing (432 egresses)    :    71 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) ││ Access/Egress routing            :  2816 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) │├ Main routing                     :   475 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) │├ Creating itineraries             :     9 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) ├┴ Transit routing total            :  3321 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) │  Routing total:                   :  3327 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) ├  Filtering itineraries            :     0 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) ├  Converting model objects         :     0 ms
10:10:48.700 DEBUG (DebugTimingAggregator.java:220) ┴  Request total                    :  3327 ms
```

The improvement is very noticeable but in my view it went only from terrible to bad.

To achieve an even better performance I think we need to dramatically lower the number of accesses being computed.

For example, if I run the same request without flex, the timing looks like this:

```
10:23:21.951 DEBUG (DebugTimingAggregator.java:220) ┌  Routing initialization           :     0 ms
10:23:21.951 DEBUG (DebugTimingAggregator.java:220) ├  Direct street routing            :     6 ms
10:23:21.951 DEBUG (DebugTimingAggregator.java:220) │┌ Creating raptor data model       :    20 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) │├ Access routing (5 accesses)      :    19 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) │├ Egress routing (432 egresses)    :    77 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) ││ Access/Egress routing            :    97 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) │├ Main routing                     :    57 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) │├ Creating itineraries             :    10 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) ├┴ Transit routing total            :   184 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) │  Routing total:                   :   190 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) ├  Filtering itineraries            :     1 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) ├  Converting model objects         :     0 ms
10:23:21.952 DEBUG (DebugTimingAggregator.java:220) ┴  Request total                    :   191 ms
```

### Unit tests
I would love to write a unit test for flex but I think I need some guidance on how to do it.

### Code style
Yes.

### Documentation
I've commented the code.

### Changelog
Yes