@leonardehrenfried 2021-11-03
I also took the liberty also bring the configuration for the vector tiles in line with what @t2gran has requested for the flex one.