@t2gran 2021-11-03
> I also took the liberty also bring the configuration for the vector tiles in line with what @t2gran has requested for the flex one.

The long term goal here is to modularize the code a bit, and hopefully this will make it a bit simpler to infer a dependency-injection framework.